requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
excel==1.0.0
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
Jinja2==2.11.1
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
Pillow==7.0.0
prettytable==0.7.2
pyecharts==1.7.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
webencodings==0.5.1
wordcloud==1.6.0
xlrd==1.2.0
